[REFACTOR/#48] OrbitSlider 캔버스 하나만 사용하도록 수정
